+++ src/microsim/schema/_emission.py
